+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ## Added
 - macOS: Add to support Mouse special key(Back, Forward)
+- win: Helper function to tell Windows to respect the users scaling settings `set_dpi_awareness`. Read the docs before using it
 
 ## Removed